move definition TestConfig class out of runtests.py (#620)
